+++ src/devtools/client/debugger/src/components/TestInfo/TestSteps.tsx
@@ -1,30 +1,22 @@
 import React, { useContext, useMemo, useState } from "react";
 
-import { setSelectedPanel, setViewMode } from "ui/actions/layout";
 import { startPlayback } from "ui/actions/timeline";
-import {
-  CanonicalRequestType,
-  RequestSummary,
-  partialRequestsToCompleteSummaries,
-} from "ui/components/NetworkMonitor/utils";
+import { partialRequestsToCompleteSummaries } from "ui/components/NetworkMonitor/utils";
 import Icon from "ui/components/shared/Icon";
 import { getEvents, getRequests } from "ui/reducers/network";
 import {
   getReporterAnnotationsForTitle,
   getReporterAnnotationsForTitleEnd,
-  getReporterAnnotationsForTitleStart,
 } from "ui/reducers/reporter";
 import { useAppDispatch, useAppSelector } from "ui/setup/hooks";
-import { AnnotatedTestStep, Annotation, TestItem, TestStep } from "ui/types";
+import { AnnotatedTestStep, TestItem, TestStep } from "ui/types";
 
+import { NetworkEvent, getDisplayedEvents } from "./NetworkEvent";
 import { TestInfoContext } from "./TestInfo";
 import { TestStepItem } from "./TestStepItem";
 
-import { setValues } from "framer-motion/types/render/utils/setters";
-import { selectAndFetchRequest } from "ui/actions/network";
-
 // const XHR_TYPE = "xhr";
-const XHR_TYPE = "text/jsx";
+export const XHR_TYPE = "text/jsx";
 
 function useGetTestSections(
   steps: TestStep[],
@@ -172,15 +164,6 @@ function TestSection({
 
   const data = partialRequestsToCompleteSummaries(requests, events, new Set());
 
-  console.log({
-    requests,
-    // filtered: requests.filter(r => r.time >= startTime && r.time < endTime),
-    events,
-    data,
-    filtered: data.filter(r => r.end && r.end >= startTime && r.end < endTime),
-    steps,
-  });
-
   return (
     <>
       {header ? <div className="py-2">{header}</div> : null}
@@ -208,52 +191,11 @@ function TestSection({
           />
           <div className="flex flex-col">
             {getDisplayedEvents(s, steps, data, startTime).map((r, i) => (
-              <NetworkEvents key={i} method={r.method} status={r.status} url={r.url} id={r.id} />
+              <NetworkEvent key={i} method={r.method} status={r.status} url={r.url} id={r.id} />
             ))}
           </div>
         </>
       ))}
     </>
-  );
-}
-
-const getDisplayedEvents = (
-  step: AnnotatedTestStep,
-  steps: AnnotatedTestStep[],
-  data: RequestSummary[],
-  startTime: number
-) => {
-  return data.filter(r => {
-    if (!r.end) {
-      return false;
-    }
-
-    const isDuringStep = (time: number, start: number, end: number) => time >= start && time < end;
-    const applicableSteps = steps.filter(s =>
-      isDuringStep(
-        r.end!,
-        startTime + s.relativeStartTime,
-        startTime + s.relativeStartTime + s.duration
-      )
-    );
-
-    return (
-      applicableSteps[applicableSteps.length - 1]?.id === step.id && r.documentType === XHR_TYPE
-    );
-  });
-};
-
-function NetworkEvents({ method, status, url, id }: { method: string; status?: number; url: string; id: string }) {
-  const dispatch = useAppDispatch();
-  const onClick = () => {
-    dispatch(setViewMode("dev"));
-    dispatch(setSelectedPanel("network"));
-    dispatch(selectAndFetchRequest(id));
-  };
-
-  return (
-    <button className="font-italic flex border-b border-themeBase-90 p-1 px-2" onClick={onClick}>
-      {method} {status} {new URL(url).pathname}
-    </button>
   );
 }
